Rename 'components/DashedRectangleTimeBarChart.vue' to 'components/index/CardsMonitoring/HospitalizedNumber/Chart.vue'
